### 2025-07-10 #8570 @editorialbot
[PRE REVIEW]: Tabbed: A Python package for reading variably structured text files at scale
**Submitting author:** @mscaudill (<a href="http://orcid.org/0000-0002-3656-9261">Matthew Caudill</a>)
**Repository:** https://github.com/mscaudill/tabbed
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
