### 2025-08-19 562cb13
[BUGFIX]: 4 bugfixes are being addressed: 1. type sniffing now uses exclusions and polling defaults removed from metadata, header and datetime formats, 2. Reader initializer now takes poll and exclude and passes them to sniffer methods for direct user access and corrects poor header determination in init, 3. Bugfix for single column data by replacing empty delimiter with carriage return, 4. type difference now ignores type differences between ints, floats and complexes since they are subtypes of each other. Sjipping test for now since we have some failures in test_sniffing to still workout [sk
[…]
