### 2025-08-20 284e8e4
[refactor]: type_difference was unclear because it was using a combination of types and row lengths to identify header and metadata, now we use separate _length_difference method for metadata location and type_difference for header location under the assumption that header rows must have the same length as polled data rows
