+++ src/tabbed/sniffing.py
@@ -147,7 +147,7 @@ class Sniffer(ReprMixin):
         >>> print(sniffer.dialect)
         SimpleDialect(';', '"', None)
         >>> # ask the sniffer to return a Header
-        >>> header = sniffer.header()
+        >>> header = sniffer.header(poll=4)
         >>> print(header)
         ... #doctest: +NORMALIZE_WHITESPACE
         Header(line=3,
@@ -296,9 +296,11 @@ class Sniffer(ReprMixin):
         """
 
         if value:
-            # python 3.11 deprecated '' for escape & quotechars
+            # python 3.11 deprecated '' for delimiter, escape & quotechars
+            delimiter = '\r' if value.delimiter == '' else value.delimiter
             escapechar = None if value.escapechar == '' else value.escapechar
             quotechar = '"' if not value.quotechar else value.quotechar
+            value.delimiter = delimiter
             value.escapechar = escapechar
             value.quotechar = quotechar
 
@@ -322,9 +324,11 @@ class Sniffer(ReprMixin):
         result = []
         delimiter = self.dialect.delimiter
 
-        # if no delimiter this is single column data delimited by newline char
-        if not delimiter:
-            return [[astring] for astring in self.sample.splitlines()]
+        # single column data uses carriage return delimiter
+        if delimiter == '\r':
+            return [
+                [astr.replace('"', '')] for astr in self.sample.splitlines()
+            ]
 
         # split sample_str on terminators, strip & split each line on delimiter
         for line in self.sample.splitlines():
@@ -357,11 +361,7 @@ class Sniffer(ReprMixin):
             next(self.infile)
 
     def _resample(self) -> None:
-        """Sample from the infile using the start, amount and skip properties.
-
-        Returns:
-            A tuple with the sample string and line numbers of the sample.
-        """
+        """Sample from infile using the start, amount and skip properties."""
 
         self._move(self.start)
         result = SimpleNamespace(indices=[], linestrs=[])
@@ -408,9 +408,6 @@ class Sniffer(ReprMixin):
             warnings.warn(msg1 + msg2)
             self._dialect = None
         else:
-            # empty str dialect disallowed by csv module
-            if result.delimiter == '':
-                result.delimiter = '\r'
             self.dialect = result
 
     # no mutation of exclude list here
@@ -496,17 +493,18 @@ class Sniffer(ReprMixin):
 
         return polled[-1], consistent
 
-    def _type_diff(
+    def _length_diff(
         self,
         poll: int,
         exclude: List[str],
-        len_requirement: bool = True,
     ) -> Tuple[int | None, List[str] | None]:
-        """Locates first row from last whose types don't match last sample row.
+        """Locates metadata by identifying the first row from the end of the
+        sample whose length does not match the length of the last poll rows.
 
-        This heuristic locates the header or metadata row by looking for type
-        changes as the sample is read from the last row backwards. It assumes
-        consistent column types in the sample.
+        This method assumes that the metadata row lengths do not match the data
+        row lengths. This can obviously be untrue but detecting the difference
+        between a header row whose length must match the number of data columns
+        from a metadata row with the same number of columns is challenging.
 
         Args:
             poll:
@@ -514,13 +512,40 @@ class Sniffer(ReprMixin):
             exclude:
                 A sequence of characters that indicate missing values. Rows
                 containing these strings will be ignored.
-            len_requirement:
-                A boolean indicating if the first row from last with a type
-                mismatch must have the same length as the last row of the
-                sample. This will be True for headers and False for metadata.
 
         Returns:
-            An integer line number and header row or a 2-tuple of Nones
+            A 2-tuple of integer line number and the metadata row if found and
+            a 2-tuple of Nones otherwise.
+        """
+
+        types, _ = self.types(poll, exclude)
[…]
