### 2025-08-20 6261361
[docs]: site updates to explain how poll and exclude parameters of reader can be used to improve header, metadata and type detection
