### 2025-09-04 7aa0091
[bugfix]: inconsistent types in sample columns triggers automatic header detection failure but this is not desired, tabbed should still look for header with inconsistent types rather than give up, this is now correct by no longer checking type consistency in _type_diff method of sniffer; this means columns containing dates and numerics or strings and numerics may incorrectly identify header but this is expected behavior; all tests pass locally
