### 2025-09-06 667d1c3
[bugfix]: ensure types method returns consistent type boolean if the types in a column are all numeric-like
