### 2025-09-10 97de637
[DOCS]: addressing reviewer jolars request to include frictionless-py in the manuscript discussion and table see https://github.com/mscaudill/tabbed/issues/2
