+++ paper/paper.bib
@@ -42,3 +42,14 @@ license = {BSD-3-Clause},
 title = {{pandas-dev/pandas: Pandas}},
 url = {https://github.com/pandas-dev/pandas}
 }
+
+@software{frictionless,
+author = {Karev, Evgeny and Camilleri, Pierre and Baptista, Vitor and Bere,
+    Georgiana and Borruso, Andrea and Desmet, Peter and Gharti, Shashi and
+    Herrmann, Augusto and Kariv, Adam and Shaw, Chris and Walsh, Paul and
+    Winfree, Lilly and Zanella Alvarenga, Edgar and Zedlitz, Jesper and
+    {Open Knowledge Foundation} and Petti, Sara},
+license = {MIT},
+title = {{frictionless: Python library for Data Packages}},
+url = {https://github.com/frictionlessdata/frictionless-py}
+}
