### 2025-09-11 mscaudill/tabbed#8 @jolars
[DOCS]: Suggestions for JOSS paper
## Major Remarks

https://github.com/mscaudill/tabbed/blob/97de63721b4d98c94a3483f5269be2e84dca5868/paper/paper.md?plain=1#L126

Wait, why is this important? Why would you want to use the Python parser in pandas when the other parsers are orders of magnitude faster? This doesn't seem like a fair comparison to me.

## Minor Remarks

https://github.com/mscaudill/tabbed/blob/97de63721b4d98c94a3483f5269be2e84dca5868/paper/paper.md?plain=1#L30

Just nitpicking, but the sentence seems to say that "automatic parsing" is frustrated.

https://github.com/mscaudill/tabbed/blob/97de63721b4d98c94a3483f5269be2e84dca5868/paper/paper.md?plain=1#L144

You alternate between "Tabbed" and "tabbed" (no initial capital letter). You might want to be consistent. Also, you use italics for clevercsv, but then shouldn't you use that for Tabbed too?

https://github.com/mscaudill/tabbed/blob/97de63721b4d98c94a3483f5269be2e84dca5868/paper/paper.md?plain=1#L147

Check the grammar here. Should it be "approximate", not "approximately"?

Ref: https://github.com/openjournals/joss-reviews/issues/8964
