### 2025-09-11 mscaudill/tabbed#10 @jolars
[DOCS]: Missing date entry for frictionless bib entry
https://github.com/mscaudill/tabbed/blob/97de63721b4d98c94a3483f5269be2e84dca5868/paper/paper.bib#L46-L55

This entry is missing a date/year + month entry. I know it's what clicking citation on the GitHub page gives you, but you should fill in the blanks.

Ref: https://github.com/openjournals/joss-reviews/issues/8964
