### 2025-09-11 7259b66
[docs, bugfix]: adding pandas publication to paper and renamed duplicate tests in test_reading caught by reviewer jolars
