+++ paper/paper.bib
@@ -43,6 +43,16 @@ title = {{pandas-dev/pandas: Pandas}},
 url = {https://github.com/pandas-dev/pandas}
 }
 
+@InProceedings{ mckinney2010,
+  author    = { {W}es {M}c{K}inney },
+  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 56 - 61 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-00a }
+}
+
 @software{frictionless,
 author = {Karev, Evgeny and Camilleri, Pierre and Baptista, Vitor and Bere,
     Georgiana and Borruso, Andrea and Desmet, Peter and Gharti, Shashi and
