### 2025-09-11 df5e4d2
[docs]: correcting pandas citations so both show up
