### 2025-09-11 df79c9c
[docs] addressing reviewer jolars tweaks to the manuscript https://github.com/mscaudill/tabbed/issues/8
