### 2025-09-11 1b0d6cd
[docs]: updating rationale for comparing tabbed's Reader performance with pandas read_csv under the python engine to satisfy https://github.com/mscaudill/tabbed/issues/8)
