### 2025-09-12 7804a89
[docs]: using zenodo's citation for pandas and frictionless software references per https://github.com/mscaudill/tabbed/issues/10
