+++ paper/paper.bib
@@ -15,14 +15,13 @@
   pages = {1799–1820}
 }
 
-@misc{tablib,
-  author = {Reitz, K.},
-  title = {tablib},
-  year = {2016},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/jazzband/tablib}},
-  commit = {8dcb87f69d2f6d501a511210387df233099b2957}
+@software{tablib,
+  author     = {Reitz, K.},
+  title      = {tablib: Pythonic Tabular Datasets},
+  year       = {2016},
+  version    = {3.8.0}
+  publisher  = {GitHub},
+  url        = {https://github.com/jazzband/tablib},
 }
 
 @misc{comma,
@@ -30,17 +29,19 @@
   title = {comma},
   year = {2020},
   publisher = {GitHub},
-  journal = {GitHub repository},
   howpublished = {\url{https://github.com/jlumbroso/comma}},
   commit = {d7669e59ae8480a84abf2ff89142ee243e5ddbd9}
 }
 
 @software{pandas,
-author = {{The pandas development team}},
-doi = {10.5281/zenodo.3509134},
-license = {BSD-3-Clause},
-title = {{pandas-dev/pandas: Pandas}},
-url = {https://github.com/pandas-dev/pandas}
+  author       = {The pandas development team},
+  title        = {pandas-dev/pandas: Pandas},
+  month        = aug,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {v2.3.2},
+  doi          = {10.5281/zenodo.16918803},
+  url          = {https://doi.org/10.5281/zenodo.16918803},
 }
 
 @InProceedings{ mckinney2010,
@@ -54,12 +55,34 @@ url = {https://github.com/pandas-dev/pandas}
 }
 
 @software{frictionless,
-author = {Karev, Evgeny and Camilleri, Pierre and Baptista, Vitor and Bere,
-    Georgiana and Borruso, Andrea and Desmet, Peter and Gharti, Shashi and
-    Herrmann, Augusto and Kariv, Adam and Shaw, Chris and Walsh, Paul and
-    Winfree, Lilly and Zanella Alvarenga, Edgar and Zedlitz, Jesper and
-    {Open Knowledge Foundation} and Petti, Sara},
-license = {MIT},
-title = {{frictionless: Python library for Data Packages}},
-url = {https://github.com/frictionlessdata/frictionless-py}
+  author       = {Karev, Evgeny and
+                  Camilleri, Pierre and
+                  Baptista, Vitor and
+                  Bere, Georgiana and
+                  Borruso, Andrea and
+                  Desmet, Peter and
+                  Gharti, Shashi and
+                  Herrmann, Augusto and
+                  Kariv, Adam and
+                  Shaw, Chris and
+                  Walsh, Paul and
+                  Winfree, Lilly and
+                  Zanella Alvarenga, Edgar and
+                  Zedlitz, Jesper and
+                  Open Knowledge Foundation and
+                  Petti, Sara},
+  title        = {frictionless: Python library for Data Packages},
+  month        = mar,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {v5.18.1},
+  doi          = {10.5281/zenodo.15085933},
+  url          = {https://doi.org/10.5281/zenodo.15085933},
+  swhid        = {swh:1:dir:c6518722a7ddee867d61bbaee04e716c814930ad
+                   ;origin=https://doi.org/10.5281/zenodo.4663759;vis
+                   it=swh:1:snp:f895c74ea3c5b2a282d2b19a38ab7560f381b
+                   cdf;anchor=swh:1:rel:a9c38748896bc29425338ea30b12c
[…]
