### 2025-09-12 1e86f6f
[docs]: using zenodo's citation for pandas and frictionless software references per https://github.com/mscaudill/tabbed/issues/10
