+++ paper/paper.bib
@@ -19,7 +19,7 @@
   author     = {Reitz, K.},
   title      = {tablib: Pythonic Tabular Datasets},
   year       = {2016},
-  version    = {3.8.0}
+  version    = {3.8.0},
   publisher  = {GitHub},
   url        = {https://github.com/jazzband/tablib},
 }
