### 2025-09-12 8b2e1bd
[docs]: updating The pandas development team authorship
