+++ paper/paper.bib
@@ -24,17 +24,17 @@
   url        = {https://github.com/jazzband/tablib},
 }
 
-@misc{comma,
+@software{comma,
   author = {Lumbroso, J.},
   title = {comma},
   year = {2020},
+  version = {0.5.4}
   publisher = {GitHub},
-  howpublished = {\url{https://github.com/jlumbroso/comma}},
-  commit = {d7669e59ae8480a84abf2ff89142ee243e5ddbd9}
+  url = {https://github.com/jlumbroso/comma},
 }
 
 @software{pandas,
-  author       = {The pandas development team},
+  author       = {{The pandas development team}},
   title        = {pandas-dev/pandas: Pandas},
   month        = aug,
   year         = 2025,
