+++ paper/paper.bib
@@ -28,7 +28,7 @@
   author = {Lumbroso, J.},
   title = {comma},
   year = {2020},
-  version = {0.5.4}
+  version = {0.5.4},
   publisher = {GitHub},
   url = {https://github.com/jlumbroso/comma},
 }
