### 2025-09-12 eec3587
[bugfix]: reviewer jolars attempted to read indexed lines that preceede the start line https://github.com/mscaudill/tabbed/issues/7, tabbed printed warnings when it should raise ValueError as this should not be allowed to occur, the protected _prime function now catches this
