### 2025-09-12 aa27a04
[docs, bugfix]: reviewer ymahlau found an error in the README example, the dialect is an attribute of the sniffer stored to a Reader instance and must be accessed accordingly
