### 2025-09-23 58bf4ab
[bugfix]: This corrects the bug found by reviwer @ymahlau regarding issue #14
