### 2025-09-23 0390960
[test]: as per @ymahlau feature request issue #16 Im including code coverage with pytest using pytest-cov and coveralls in testing workflow
