+++ pyproject.toml
@@ -35,6 +35,7 @@ requires-python = ">=3.12"
 dev = [
     "ipython",
     "pytest",
+    "pytest-cov",
     "mypy",
     "pylint",
     "isort",
@@ -51,7 +52,7 @@ dev = [
 ]
 linting = ["codespell", "isort", "pylint", "black"]
 typing = ["mypy"]
-testing = ["pytest"]
+testing = ["pytest", "pytest-cov"]
 
 [project.urls]
 Homepage = "https://github.com/mscaudill/tabbed"
