### 2025-09-23 996ce31
[test]: as per @ymahlau feature request issue #16 Im including code coverage with pytest using pytest-cov and coveralls in testing workflow
