### 2025-09-23 400dbaf
[test]: ignore a coveralls failure to publish since site may be down with 504 or 500 errors
