### 2025-09-23 f574e6b
[docs]: adding readme badge for code coverage per feature request #16
