### 2025-09-25 93b8cce
[test]: remove coverage of mixins module used for printing instances in tabbed
