### 2025-09-26 9d72eb8
[feature, docs]: addresses reviewer @ymahlau request for plotting script issue #18 and addresses speed changes that occurred as the ',' decimal mark support was added as requested in issue #17
