+++ paper/paper.md
@@ -128,8 +128,8 @@ like pandas `python` engine or tabbed's `Reader` class.  The speed test was
 conducted on a DSV file with all floats and a DSV file with floats and datetime
 instances. The file size in each case was 10 columns by 100,000 rows. The left
 panel of \autoref{fig: figure2} shows that tabbed is slower on both the float
-and mixed type files by approximately 30%.  Nevertheless, users can expect to
-read millions of cells from a DSV file in just a few seconds with tabbed.
+and mixed type files.  Nevertheless, users can expect to read millions of cells
+from a DSV file in just a few seconds with tabbed.
 
 ![Tabbed and pandas read speed comparison. Left: Comparison of number of cells
 casted per second between tabbed and pandas for DSV files composed of floats
