+++ paper/plots.py
@@ -1,0 +1,291 @@
+import csv
+import datetime
+from itertools import batched
+from itertools import chain
+import pathlib
+import random
+import time
+
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+
+import tabbed
+
+ROWS=100e3
+COLS=10
+
+def write_floatcsv(path=None, name='floatcsv.txt', seed=0, rows=ROWS, cols=COLS):
+    """Writes a text file of stringed float values to path.
+
+    Args:
+        path:
+            A string or path instance to a directory where this float csv will
+            be written to.
+        name:
+            The name of the file to write.
+        seed:
+            A seed for Python's random number generator.
+        columns:
+            The number of columns to write.
+        rows:
+            The number of data rows to write.
+    """
+
+    # header & data generation
+    header = [f'Column_{i}' for i in range(cols)]
+    rgen = random.Random()
+    rgen.seed(seed)
+    data = [rgen.random() for _ in range(int(rows * cols))]
+    data = batched(data, cols)
+    # create path and write
+    path = pathlib.Path.cwd() if not path else Path(path)
+    fp = path / name
+    with open(fp, 'w+') as outfile:
+        writer = csv.writer(outfile)
+        writer.writerow(header)
+        for idx, row in enumerate(data):
+            print(f'Writing float data: {idx/rows * 100:.0f}% complete',
+                  end='\r', flush=True)
+            writer.writerow(row)
+
+    print('write complete')
+
+
+def write_mixedcsv(path=None, name='mixedcsv.txt', seed=0, rows=ROWS, cols=COLS):
+    """Writes a text file to path containing floats and datetimes.
+
+    Args:
+        path:
+            A string or path instance to a directory where this mixed csv will
+            be written to.
+        name:
+            The name of the file to write.
+        seed:
+            A seed for Python's random number generator.
+        columns:
+            The number of columns to write.
+        rows:
+            The number of data rows to write.
+    """
+
+    # header & data generation
+    header = [f'Column_{i}' for i in range(cols)]
+    rgen = random.Random()
+    rgen.seed(seed)
+    data = [rgen.random() for _ in range(int(rows * cols))]
+    data = [list(tup) for tup in batched(data, cols)]
+
+    # set the second column to be a date
+    for row in data:
+        row[2] = '9/22/2023'
+
+    # create path and write
+    path = pathlib.Path.cwd() if not path else Path(path)
+    fp = path / name
+    with open(fp, 'w+') as outfile:
+        writer = csv.writer(outfile)
+        writer.writerow(header)
+        for idx, row in enumerate(data):
+            print(f'Writing mixed data: {idx/rows * 100:.0f}% complete',
+                  end='\r', flush=True)
+            writer.writerow(row)
+
+    print('write complete')
+
+
+def time_tabbed(fp, trials):
+    """Computes the number of cells per second tabbed reads for each trial in
+    trials.
+
+    Args:
+        fp:
+            Path to a text file to read and parse.
+        trials:
+            Number of times to read the file
+
+    Returns:
+        A list of speeds in cells per sec units one per trial in trials.
+    """
+
+    speeds = []
+    for trial in range(trials):
+        infile = open(fp, 'r')
+        reader = tabbed.reading.Reader(infile)
+        gen = reader.read(chunksize=int(10e3))
+        t0 = time.perf_counter()
+        data = list(chain.from_iterable(gen))
+        elapsed = time.perf_counter() - t0
[…]
