### 2025-09-26 fb633bc
[feature, typing, test]: this commit addresses @ymahlau issue #17 requesting support for floats and complex numbers using ',' as the decimal mark, it also addresses @jolars issue #6 to use generic types in place of typing module imports
