### 2025-09-30 mscaudill/tabbed#19 @jolars
[Quest]: Metadata parsing
### Has this question been asked before?

- [x] I have searched the existing issues for this question.

### Question

If I change `annotations.txt` to:

```
Experiment ID	Experiment
Animal ID	Animal
Researcher	Test
Directory path

-----------------------------------------------------------------------------

Number	Start Time	End Time	Time From Start	Channel	Annotation
0	02/09/22 09:17:38.948	02/09/22 09:17:38.948	0.0000	ALL	Started Recording
1	02/09/22 09:37:00.000	02/09/22 09:37:00.000	1161.0520	ALL	start
2	02/09/22 09:37:00.000	02/09/22 09:37:08.784	1161.0520	ALL	exploring
3	02/09/22 09:37:08.784	02/09/22 09:37:13.897	1169.8360	ALL	grooming
```

etc

Then parsing the metadata includes the dashes:

```python
>>> reader.metadata()
MetaData(lines=(0, 7), string='Experiment ID\tExperiment\nAnimal ID\tAnimal\nResearcher\tTest\nDirectory p
ath\n\n-----------------------------------------------------------------------------\n')
>>>
```

Is this intentional?

Ref: https://github.com/openjournals/joss-reviews/issues/8964

### Code of Conduct

- [x] I agree to follow this project's Code of Conduct
