### 2025-09-30 mscaudill/tabbed#19 @jolars
Ah, hm, yes I guess this is intentional. Sorry for the noise. For some reason I thought that tabbed would try to parse the metadata, but I understand that's not the case.
