### 2025-09-30 b72158f
[docs]: adding additional clarification per issue #21 regarding proper setting of the poll number for small files
