### 2025-09-30 36ac852
[docs]: per issue #6 replacing typing.Optional with pipe, typing.Callable with abc.Callable and typing.Type with type generics, update toml to have pylint check for deprecated typing imports in future
