### 2025-09-30 21f85f9
[docs]: adding annotation for decimal arg in sniffing.Sniffer
