### 2025-10-02 78c7724
[bugfix]: encountered a file where some lines end in delimiters and others do not, now normalize sniffer rows to same length for consistent type and datatime format detection
