### 2025-10-08 d4b95f8
[bugfix]: empty files now raise IOError during Reader initialization per issue #23, for testing this requires change from TemporaryFile to NamedTemporaryFile fixtures so the name is available for file size stat, also NamedTempFiles are not iterators so the infile in sniffers move method reader's _prime method must be explicitly converted to iterator type
