### 2025-10-08 4c91dc9
[refactor]: removing line normalization after rstrip since this impacts the metadata and header rows too
