+++ src/tabbed/sniffing.py
@@ -345,13 +345,6 @@ class Sniffer(ReprMixin):
             row = [astring.replace('"', '') for astring in row]
             result.append(row)
 
-        # if some lines end in delimiters and others do not
-        # we need to normalize after rstripping
-        expected = len(max(result, key=len))
-        for row in result:
-            if len(row) < expected:
-                row.append('')
-
         return result
 
     def _move(self, line: int) -> None:
@@ -543,7 +536,6 @@ class Sniffer(ReprMixin):
 
         types, _ = self.types(poll, exclude)
         for idx, row in reversed(list(zip(self.lines, self.rows))):
-
             if len(row) != len(types):
                 return idx, row
 
