### 2025-10-08 94153cd
[bugfix]: per issue #21, tabbed now reduces the poll amount to the last row of the data section if the poll amount exceeds the number of sampled
