### 2025-10-20 e4b8579
[feature]: adding YYYY-MM-DD per issue #25
