### 2025-11-07 d900c46
[bugfix]: As per issue #24, if a small file is polled and the polling includes header and metadata then tabbed should warn of inconsistent types but still attempt to locate header and/or metadata
