### 2025-11-07 bdcd654
bump version 1.1.1 -> 1.2.0
