+++ src/tabbed/__init__.py
@@ -1,6 +1,6 @@
 """Tabbed's version"""
 
-__version__ = "1.1.1"
+__version__ = "1.2.0"
 
 from . import reading, sniffing, tabbing, utils
 
