+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "tabbed"
-version = "1.1.1"
+version = "1.2.0"
 description = """
 An iterative reader of irregular text files
 """
@@ -66,7 +66,7 @@ where = ["src"]
 
 # bumpver configuration
 [tool.bumpver]
-current_version = "1.1.1"
+current_version = "1.2.0"
 version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
 commit_message = 'bump version {old_version} -> {new_version}'
 commit = true
