### 2025-11-10 de179e0
[docs]: removing newline character to try to get paper to compile XML metadata properly with JOSS.
