+++ paper/paper.md
@@ -98,9 +98,10 @@ a first-in-first-out (FIFO) data structure with O(1) time complexity allowing
 tabbed to linearly scale to large files. 
 
 # Comparison
-Tablib [@tablib], comma [@comma], pandas [@pandas; @mckinney2010] and frictionless-py
-[@frictionless] are popular alternative packages to tabbed. \autoref{tbl:
-table1} compares their respective features. Pandas `read_csv` and Frictionless'
+Tablib [@tablib], comma [@comma], pandas [@pandas; @mckinney2010] and
+frictionless-py [@frictionless] are popular alternative packages to tabbed.
+\autoref{tbl: table1} compares their respective features. Pandas `read_csv` and
+Frictionless'
 `extract` functions most closely match the available features in tabbed. Both
 support broad type casting and iterative reading of large files. However, both
 require specifying the header row if metadata is written to the file. This per
@@ -132,13 +133,12 @@ and mixed type files.  Nevertheless, users can expect to read millions of cells
 from a DSV file in just a few seconds with tabbed.
 
 ![Tabbed and pandas read speed comparison. Left: Comparison of number of cells
-casted per second between tabbed and pandas for DSV files composed of floats
-or mixed types. The conversion engine for pandas was chosen to be
-"python". Right: Tabbed's read speed as a function of the number of filters
-applied to each row during reading. In both panels black circles and error bars
-are the mean and standard deviation across 30 trials. These comparisons were
-carried out on a single 2.4 GHz Intel Core i5-6300U processor. \label{fig:
-figure2}](figure2.png)
+casted per second between tabbed and pandas for DSV files composed of floats or
+mixed types. The conversion engine for pandas was chosen to be "python". Right:
+Tabbed's read speed as a function of the number of filters applied to each row
+during reading. In both panels black circles and error bars are the mean and
+standard deviation across 30 trials. These comparisons were carried out on
+a single 2.4 GHz Intel Core i5-6300U processor \label{fig: figure2}](figure2.png).
 
 To further understand tabbed's performance, we tested how row filtering impacts
 read speeds.  The right panel of \autoref{fig: figure2} shows the read speed
