### 2025-11-10 e14f75a
[docs, bugfix]: unclosed fence in paper
