### 2025-11-10 c640ce6
[docs, bugfix]: removing period at end of figure2 label, it should not be there
