### 2025-11-10 #8964 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
